Add threaded requests
This implementation is certainly far from perfect and has some issues,
but it's good enough for now. The main issue is that multiple requests
at the same time is not possible. I'm not sure why this is. I believe
it may be possible with CurlMulti, which I'll look into at a later time.

The original, blocking method is now called getBlocking.