Remove CPR dependency and move to QSslSocket
To explain reason for not just using QNetworkAccessManager again:
QNetworkAccessManager adds a couple of its own headers to every single
request it makes. One of these is rejected by Innertube and it will flat
out not give you responses. However, QSslSocket (and CPR as well) does
not do this, so it's what has to be used.